requirements.txt


accelerate==0.30.1
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
backoff==2.2.1
bcrypt==4.1.3
beautifulsoup4==4.12.3
bitsandbytes==0.43.1
blinker==1.8.2
brotli==1.1.0
certifi==2023.7.22
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.2
chromadb==0.4.7
click==8.1.7
colorama==0.4.6
colorclass==2.2.2
coloredlogs==15.0.1
compressed-rtf==1.0.6
comtypes==1.4.2
contourpy==1.2.1
cryptography==42.0.7
cssselect2==0.7.0
cycler==0.10.0
dataclasses-json==0.5.14
decorator==4.4.2
defusedxml==0.7.1
deprecated==1.2.14
dirtyjson==1.0.8
diskcache==5.6.3
distro==1.9.0
dnspython==2.6.1
doctr==1.9.0
docx2txt==0.8
easygui==0.98.3
easyocr==1.7.1
ebcdic==1.1.1
emoji==2.11.1
extract-msg==0.45.0
fastapi==0.99.1
filelock==3.14.0
filetype==1.2.0
fire==0.6.0
flask==3.0.3
flatbuffers==24.3.25
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.3.1
googletrans==3.0.0
gpt4all==1.0.8
greenlet==3.0.3
gtts==2.5.1
h11==0.14.0
h2==3.2.0
h5py==3.11.0
hpack==3.0.0
hstspreload==2024.5.1
html5lib==1.1
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
hub-sdk==0.0.8
huggingface-hub==0.23.0
humanfriendly==10.0
hyperframe==5.2.0
idna==2.10
imageio==2.34.1
imageio-ffmpeg==0.4.9
imapclient==2.3.1
importlib-metadata==7.1.0
importlib-resources==6.4.0
intel-openmp==2021.4.0
itsdangerous==2.2.0
jinja2==3.1.4
joblib==1.4.0
kiwisolver==1.4.5
langchain==0.0.274
langdetect==1.0.9
langsmith==0.0.92
lark==1.1.8
lazy-loader==0.4
libretranslatepy==2.1.1
llama-cpp-python==0.1.81
llama-index==0.10.30
llama-index-agent-openai==0.2.2
llama-index-cli==0.1.12
llama-index-core==0.10.30
llama-index-embeddings-openai==0.1.8
llama-index-indices-managed-llama-cloud==0.1.5
llama-index-legacy==0.9.48
llama-index-llms-openai==0.1.16
llama-index-multi-modal-llms-openai==0.1.5
llama-index-program-openai==0.1.5
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.19
llama-index-readers-llama-parse==0.1.4
llama-parse==0.4.1
llamaindex-py-client==0.1.18
lxml==5.2.1
markupsafe==2.1.5
marshmallow==3.21.1
matplotlib==3.8.4
mkl==2021.4.0
monotonic==1.6
moviepy==1.0.3
mplcursors==0.5.3
mpmath==1.3.0
msoffcrypto-tool==5.4.0
multidict==6.0.5
mypy-extensions==1.0.0
mysql-connector-python==8.4.0
nest-asyncio==1.6.0
networkx==3.3
ninja==1.11.1.1
nltk==3.8.1
numexpr==2.10.0
numpy==1.26.4
olefile==0.46
oletools==0.60.1
ollama==0.1.9
onnxruntime==1.17.3
openai==1.23.2
opencv-python==4.8.1.78
opencv-python-headless==4.8.0.74
overrides==7.7.0
packaging==24.0
pandas==2.2.2
pandoc==2.3
pcodedmp==1.2.6
pillow==10.3.0
plumbum==1.8.3
ply==3.11
posthog==3.5.0
proglog==0.1.10
protobuf==3.20.3
psutil==5.9.8
pulsar-client==3.5.0
py-cpuinfo==9.0.0
pybboxes==0.1.6
pyclipper==1.3.0.post5
pycparser==2.22
pydantic==1.10.15
pydantic-core==2.18.1
pydub==0.25.1
pydyf==0.10.0
pymongo==4.7.2
pymupdf==1.23.5
pymupdfb==1.23.5
pyodbc==5.1.0
pypandoc==1.11
pyparsing==2.4.7
pypdf==4.2.0
pypdfium2==4.30.0
pyphen==0.15.0
pypika==0.48.9
pypiwin32==223
pypyodbc==1.3.6
pyreadline3==3.4.1
pytesseract==0.3.10
python-bidi==0.4.2
python-dateutil==2.9.0.post0
python-doctr==0.8.1
python-docx==1.1.2
python-dotenv==1.0.0
python-facepp==0.1.0a5
python-magic==0.4.27
python-multipart==0.0.9
python-pptx==0.6.23
pyttsx3==2.90
pytz==2024.1
pywin32==306
pyyaml==6.0.1
rapidfuzz==3.9.1
red-black-tree-mod==1.20
regex==2024.4.16
requests==2.31.0
requests-toolbelt==1.0.0
rfc3986==1.5.0
roboflow==1.1.29
rtfde==0.1.1
safetensors==0.4.3
sahi==0.11.16
scikit-image==0.23.2
scikit-learn==1.4.2
scipy==1.13.0
seaborn==0.13.2
sentence-transformers==2.2.2
sentencepiece==0.2.0
shapely==2.0.4
six==1.16.0
sniffio==1.3.1
soundfile==0.12.1
soupsieve==2.5
speechrecognition==3.10.4
sqlalchemy==2.0.29
starlette==0.27.0
striprtf==0.0.26
supervision==0.20.0
sympy==1.12
tabulate==0.9.0
tbb==2021.12.0
tenacity==8.2.3
termcolor==2.4.0
terminaltables==3.1.10
thop==0.1.1-2209072238
threadpoolctl==3.5.0
tifffile==2024.5.10
tiktoken==0.6.0
tinycss2==1.3.0
tokenizers==0.19.1
torch==2.3.0
torchvision==0.18.0
tqdm==4.66.1
transformers==4.40.2
translate==3.6.1
typing-extensions==4.11.0
typing-inspect==0.9.0
tzdata==2024.1
tzlocal==5.2
ultralytics==8.0.239
ultralyticsplus==0.1.0
unidecode==1.3.8
unstructured==0.10.8
urllib3==2.0.4
uvicorn==0.29.0
watchfiles==0.21.0
weasyprint==62.1
webencodings==0.5.1
websockets==12.0
werkzeug==3.0.3
